+++ frameworks/react/src/components/combobox/combobox-item-group-label.tsx
@@ -1,19 +1,19 @@
-import type { ItemGroupLabelProps } from '@zag-js/combobox'
 import { mergeProps } from '@zag-js/react'
 import { forwardRef } from 'react'
-import type { Assign } from '../../types'
-import { createSplitProps } from '../../utils/create-split-props'
 import { type HTMLArkProps, ark } from '../factory'
 import { useComboboxContext } from './use-combobox-context'
+import { useComboboxItemGroupPropsContext } from './use-combobox-item-group-props-context'
 
-export interface ComboboxItemGroupLabelProps
-  extends Assign<HTMLArkProps<'div'>, ItemGroupLabelProps> {}
+export interface ComboboxItemGroupLabelProps extends HTMLArkProps<'div'> {}
 
 export const ComboboxItemGroupLabel = forwardRef<HTMLDivElement, ComboboxItemGroupLabelProps>(
   (props, ref) => {
-    const [optionProps, localProps] = createSplitProps<ItemGroupLabelProps>()(props, ['htmlFor'])
     const combobox = useComboboxContext()
-    const mergedProps = mergeProps(combobox.getItemGroupLabelProps(optionProps), localProps)
+    const itemGroupProps = useComboboxItemGroupPropsContext()
+    const mergedProps = mergeProps(
+      combobox.getItemGroupLabelProps({ htmlFor: itemGroupProps.id }),
+      props,
+    )
 
     return <ark.div {...mergedProps} ref={ref} />
   },