+++ frameworks/solid/src/components/combobox/use-combobox-item-group-props-context.ts
@@ -1,0 +1,10 @@
+import type { ItemGroupProps } from '@zag-js/combobox'
+import { createContext } from '../../utils/create-context'
+
+export interface UseComboboxItemGroupPropsContext extends ItemGroupProps {}
+
+export const [ComboboxItemGroupPropsProvider, useComboboxItemGroupPropsContext] =
+  createContext<ItemGroupProps>({
+    hookName: 'useComboboxItemGroupPropsContext',
+    providerName: '<ComboboxItemGroupPropsProvider />',
+  })