+++ frameworks/vue/src/components/combobox/combobox-item-group.tsx
@@ -1,17 +1,20 @@
-import type { ItemGroupProps } from '@zag-js/combobox'
-import { type PropType, defineComponent } from 'vue'
-import type { Assign } from '../../types'
+import { type PropType, computed, defineComponent } from 'vue'
+import { useId } from '../../utils'
 import { type HTMLArkProps, ark } from '../factory'
 import { useComboboxContext } from './use-combobox-context'
+import { ComboboxItemGroupPropsProvider } from './use-combobox-item-group-props-context'
 
-export interface ComboboxItemGroupProps extends Assign<HTMLArkProps<'div'>, ItemGroupProps> {}
+export interface ComboboxItemGroupProps extends HTMLArkProps<'div'> {}
 
 export const ComboboxItemGroup = defineComponent<ComboboxItemGroupProps>(
   (props, { slots, attrs }) => {
-    const api = useComboboxContext()
+    const combobox = useComboboxContext()
+    const id = useId()
+    const itemGroupProps = computed(() => ({ id: props.id ? props.id : id.value }))
 
+    ComboboxItemGroupPropsProvider(itemGroupProps.value)
     return () => (
-      <ark.div {...api.value.getItemGroupProps(props)} {...attrs}>
+      <ark.div {...combobox.value.getItemGroupProps(itemGroupProps.value)} {...attrs}>
         {slots.default?.()}
       </ark.div>
     )
@@ -21,7 +24,7 @@ export const ComboboxItemGroup = defineComponent<ComboboxItemGroupProps>(
     props: {
       id: {
         type: String as PropType<ComboboxItemGroupProps['id']>,
-        required: true,
+        required: false,
       },
     },
   },